wearable_detection_backend/requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.5.6
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bayesian-optimization==1.2.0
BayesianOptimization==0.0.0
bert4keras==0.10.6
bleach==3.2.1
blessings==1.7
blis==0.7.4
boto3==1.18.15
botocore==1.21.15
cached-property==1.5.2
cachetools==4.1.1
catalogue==2.0.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
Django==3.2.8
django-simpleui==2021.5.11
djangorestframework==3.12.4
dtaidistance==2.3.2
dtw-python==1.1.10
einops==0.3.2
entrypoints==0.3
fasttext==0.9.2
filelock==3.0.12
ftfy==5.9
future==0.18.2
gast==0.3.3
gensim==3.2.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.34.0
h5py==2.10.0
heartpy==1.2.7
idna==2.10
imageio==2.9.0
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==3.1.1
importlib-resources==5.1.2
ipdb==0.13.9
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.8.0
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kaggle==1.5.12
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lightgbm==3.2.0
llvmlite==0.36.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.6.2
notebook==6.1.5
numba==0.53.1
numpy==1.21.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
portalocker==2.0.0
preshed==3.0.5
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.0
pyCompare==1.5.1
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.2
pylint==2.8.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==5.0.2
pytorch-metric-learning==0.9.99
pytorch-pretrained-bert==0.6.2
pytorchtools==0.0.2
pyts==0.11.0
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.5.0
sacrebleu==1.5.1
sacremoses==0.0.43
scikit-image==0.18.1
scikit-learn==0.24.2
scikit-multilearn==0.2.0
scikit-optimize==0.8.1
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
sklearn==0.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
spacy-pkuseg==0.0.28
sqlparse==0.4.2
srsly==2.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow-addons==0.13.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-hub==0.12.0
tensorflow-metadata==0.28.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
thinc==8.0.3
threadpoolctl==2.1.0
tifffile==2021.3.17
tokenizers==0.9.4
toml==0.10.2
torch==1.9.0
torch-tb-profiler==0.1.0
torchcontrib==0.0.2
torchsummary==1.5.1
torchtext==0.9.1
torchvision==0.9.1
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
transformers==4.2.2
tslearn==0.5.2
typed-ast==1.4.3
typeguard==2.12.1
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wfdb==3.4.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
